Move examples
