+++ examples/.gitignore
@@ -1,2 +1,3 @@
 pokemon-service-client/
 pokemon-service-server-sdk/
+Cargo.lock
