+++ examples/pokemon-service-common/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "pokemon-service-common"
+version = "0.1.0"
+edition = "2021"
+publish = false
+authors = ["Smithy-rs Server Team <smithy-rs-server@amazon.com>"]
+description = "A smithy Rust service to retrieve information about Pokémon."
+
+[dependencies]
+async-stream = "0.3"
+http = "0.2.9"
+rand = "0.8"
+tower = "0.4"
+tracing = "0.1"
+tracing-subscriber = { version = "0.3.15", features = ["env-filter", "json"] }
+tokio = { version = "1", default-features = false, features = ["time"] }
+
+# Local paths
+aws-smithy-http = { path = "../../rust-runtime/aws-smithy-http" }
+aws-smithy-http-server = { path = "../../rust-runtime/aws-smithy-http-server" }
+pokemon-service-client = { path = "../pokemon-service-client" }
+pokemon-service-server-sdk = { path = "../pokemon-service-server-sdk" }
