+++ examples/pokemon-service-common/src/plugin.rs
