+++ examples/pokemon-service-common/tests/plugins_execution_order.rs
@@ -2,13 +2,13 @@
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
+
 use aws_smithy_http::body::SdkBody;
 use aws_smithy_http_server::operation::Operation;
 use aws_smithy_http_server::plugin::{Plugin, PluginPipeline};
-use hyper::http;
-use pokemon_service::do_nothing;
 use pokemon_service_client::operation::do_nothing::DoNothingInput;
 use pokemon_service_client::Config;
+use pokemon_service_common::do_nothing;
 use std::ops::Deref;
 use std::sync::Arc;
 use std::sync::Mutex;
@@ -59,7 +59,10 @@ struct SentinelPlugin {
 
 impl SentinelPlugin {
     pub fn new(name: &'static str, output: Arc<Mutex<Vec<&'static str>>>) -> Self {
-        Self { name, output: output }
+        Self {
+            name,
+            output: output,
+        }
     }
 }
