+++ examples/pokemon-service-lambda/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "pokemon-service-lambda"
+version = "0.1.0"
+edition = "2021"
+publish = false
+authors = ["Smithy-rs Server Team <smithy-rs-server@amazon.com>"]
+description = "A smithy Rust service to retrieve information about Pokémon."
+
+[dependencies]
+async-stream = "0.3"
+clap = { version = "~3.2.1", features = ["derive"] }
+hyper = {version = "0.14.12", features = ["server"] }
+tokio = "1.20.1"
+tower = "0.4"
+tracing = "0.1"
+
+# This dependency is only required for the `pokemon-service-lambda` program.
+lambda_http = "0.7.1"
+
+# Local paths
+aws-smithy-http-server = { path = "../../rust-runtime/aws-smithy-http-server", features = ["aws-lambda"] }
+pokemon-service-server-sdk = { path = "../pokemon-service-server-sdk/" }
+pokemon-service-common = { path = "../pokemon-service-common/" }
