+++ examples/pokemon-service-lambda/tests/fixtures/example-apigw-request.json
