+++ examples/pokemon-service-tls/Cargo.toml
@@ -1,0 +1,38 @@
+[package]
+name = "pokemon-service-tls"
+version = "0.1.0"
+edition = "2021"
+publish = false
+authors = ["Smithy-rs Server Team <smithy-rs-server@amazon.com>"]
+description = "A smithy Rust service to retrieve information about Pokémon."
+
+[dependencies]
+clap = { version = "4.1.11", features = ["derive"] }
+hyper = { version = "0.14.12", features = ["server"] }
+tokio = "1.20.1"
+
+# These dependencies are only required for the `pokemon-service-tls` program.
+tls-listener = { version = "0.5.1", features = ["rustls", "hyper-h2"] }
+tokio-rustls = "0.23.4"
+rustls-pemfile = "1.0.1"
+futures-util = { version = "0.3.16", default-features = false }
+
+# Local paths
+aws-smithy-http-server = { path = "../../rust-runtime/aws-smithy-http-server" }
+pokemon-service-server-sdk = { path = "../pokemon-service-server-sdk/" }
+pokemon-service-common = { path = "../pokemon-service-common/" }
+
+[dev-dependencies]
+assert_cmd = "2.0"
+home = "0.5"
+http = "0.2.9"
+serial_test = "0.7.0"
+
+# This dependency is only required for testing the `pokemon-service-tls` program.
+hyper-rustls = { version = "0.23.0", features = ["http2"] }
+
+# Local paths
+aws-smithy-client = { path = "../../rust-runtime/aws-smithy-client/", features = ["rustls"] }
+aws-smithy-http = { path = "../../rust-runtime/aws-smithy-http/" }
+aws-smithy-types = { path = "../../rust-runtime/aws-smithy-types/" }
+pokemon-service-client = { path = "../pokemon-service-client/" }
