+++ examples/pokemon-service/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "pokemon-service"
+version = "0.1.0"
+edition = "2021"
+publish = false
+authors = ["Smithy-rs Server Team <smithy-rs-server@amazon.com>"]
+description = "A smithy Rust service to retrieve information about Pokémon."
+
+[dependencies]
+clap = { version = "4.1.11", features = ["derive"] }
+hyper = {version = "0.14.12", features = ["server"] }
+tokio = "1.20.1"
+tracing = "0.1"
+
+# Local paths
+aws-smithy-http-server = { path = "../../rust-runtime/aws-smithy-http-server", features = ["request-id"] }
+pokemon-service-server-sdk = { path = "../pokemon-service-server-sdk/" }
+pokemon-service-common = { path = "../pokemon-service-common/" }
+
+[dev-dependencies]
+assert_cmd = "2.0"
+async-stream = "0.3"
+home = "0.5"
+rand = "0.8.5"
+serial_test = "0.7.0"
+
+# This dependency is only required for testing the `pokemon-service-tls` program.
+hyper-rustls = { version = "0.23.0", features = ["http2"] }
+
+# Local paths
+aws-smithy-client = { path = "../../rust-runtime/aws-smithy-client/", features = ["rustls"] }
+aws-smithy-http = { path = "../../rust-runtime/aws-smithy-http/" }
+aws-smithy-types = { path = "../../rust-runtime/aws-smithy-types/" }
+pokemon-service-client = { path = "../pokemon-service-client/" }
