+++ examples/pokemon-service/tests/event_streaming.rs
@@ -3,24 +3,21 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-// Files here are for running integration tests.
-// These tests only have access to your crate's public API.
-// See: https://doc.rust-lang.org/book/ch11-03-test-organization.html#integration-tests
-
-use crate::helpers::{client, client_http2_only, PokemonService};
+pub mod common;
 
 use async_stream::stream;
-use aws_smithy_types::error::display::DisplayErrorContext;
-use pokemon_service_client::{
-    error::SdkError,
-    operation::get_storage::GetStorageError,
-    types::error::{AttemptCapturingPokemonEventError, MasterBallUnsuccessful, StorageAccessNotAuthorized},
-    types::{AttemptCapturingPokemonEvent, CapturingEvent, CapturingPayload},
-};
 use rand::Rng;
 use serial_test::serial;
 
-mod helpers;
+use pokemon_service_client::types::{
+    error::{AttemptCapturingPokemonEventError, MasterBallUnsuccessful},
+    AttemptCapturingPokemonEvent, CapturingEvent, CapturingPayload,
+};
+
+fn get_pokemon_to_capture() -> String {
+    let pokemons = vec!["Charizard", "Pikachu", "Regieleki"];
+    pokemons[rand::thread_rng().gen_range(0..pokemons.len())].to_string()
+}
 
 fn get_pokeball() -> String {
     let random = rand::thread_rng().gen_range(0..100);
@@ -36,82 +33,12 @@ fn get_pokeball() -> String {
     pokeball.to_string()
 }
 
-fn get_pokemon_to_capture() -> String {
-    let pokemons = vec!["Charizard", "Pikachu", "Regieleki"];
-    pokemons[rand::thread_rng().gen_range(0..pokemons.len())].to_string()
-}
-
-#[tokio::test]
-#[serial]
-async fn test_check_health() {
-    let _program = PokemonService::run().await;
-
-    let _check_health = client().check_health().send().await.unwrap();
-}
-
-#[tokio::test]
-#[serial]
-async fn test_check_health_http2() {
-    // Make sure our server can serve http2
-    let _program = PokemonService::run_https().await;
-    let _check_health = client_http2_only().check_health().send().await.unwrap();
-}
-
-#[tokio::test]
-#[serial]
-async fn simple_integration_test() {
-    let _program = PokemonService::run().await;
-
-    let service_statistics_out = client().get_server_statistics().send().await.unwrap();
-    assert_eq!(0, service_statistics_out.calls_count.unwrap());
-
-    let pokemon_species_output = client().get_pokemon_species().name("pikachu").send().await.unwrap();
-    assert_eq!("pikachu", pokemon_species_output.name().unwrap());
-
-    let service_statistics_out = client().get_server_statistics().send().await.unwrap();
-    assert_eq!(1, service_statistics_out.calls_count.unwrap());
-
-    let storage_err = client().get_storage().user("ash").passcode("pikachu321").send().await;
-    let has_not_authorized_error = if let Err(SdkError::ServiceError(context)) = storage_err {
-        matches!(
-            context.err(),
-            GetStorageError::StorageAccessNotAuthorized(StorageAccessNotAuthorized { .. }),
-        )
-    } else {
-        false
-    };
-    assert!(has_not_authorized_error, "expected NotAuthorized error");
-
-    let storage_out = client()
-        .get_storage()
-        .user("ash")
-        .passcode("pikachu123")
-        .send()
-        .await
-        .unwrap();
-    assert_eq!(Some(vec![]), storage_out.collection);
-
-    let pokemon_species_error = client()
-        .get_pokemon_species()
-        .name("some_pokémon")
-        .send()
-        .await
-        .unwrap_err();
-    let message = DisplayErrorContext(pokemon_species_error).to_string();
-    let expected = r#"ResourceNotFoundError [ResourceNotFoundException]: Requested Pokémon not available"#;
-    assert!(
-        message.contains(expected),
-        "expected '{message}' to contain '{expected}'"
-    );
-
-    let service_statistics_out = client().get_server_statistics().send().await.unwrap();
-    assert_eq!(2, service_statistics_out.calls_count.unwrap());
-}
-
 #[tokio::test]
 #[serial]
 async fn event_stream_test() {
-    let _program = PokemonService::run().await;
+    let _child = common::run_server().await;
+
+    let client = common::client();
 
     let mut team = vec![];
     let input_stream = stream! {
@@ -145,7 +72,7 @@ async fn event_stream_test() {
     };
 
     // Throw many!
-    let mut output = client()
+    let mut output = common::client()
         .capture_pokemon()
         .region("Kanto")
         .events(input_stream.into())
@@ -156,7 +83,13 @@ async fn event_stream_test() {
         match output.events.recv().await {
             Ok(Some(capture)) => {
                 let pokemon = capture.as_event().unwrap().name.as_ref().unwrap().clone();
-                let pokedex = capture.as_event().unwrap().pokedex_update.as_ref().unwrap().clone();
+                let pokedex = capture
+                    .as_event()
+                    .unwrap()
+                    .pokedex_update
+                    .as_ref()
+                    .unwrap()
+                    .clone();
                 let shiny = if *capture.as_event().unwrap().shiny.as_ref().unwrap() {
                     ""
                 } else {
@@ -190,7 +123,7 @@ async fn event_stream_test() {
                 .build()
             ))
         };
-        let mut output = client()
+        let mut output = client
             .capture_pokemon()
             .region("Kanto")
             .events(input_stream.into())
@@ -200,7 +133,13 @@ async fn event_stream_test() {
         match output.events.recv().await {
             Ok(Some(capture)) => {
                 let pokemon = capture.as_event().unwrap().name.as_ref().unwrap().clone();
-                let pokedex = capture.as_event().unwrap().pokedex_update.as_ref().unwrap().clone();
+                let pokedex = capture
+                    .as_event()
+                    .unwrap()
+                    .pokedex_update
+                    .as_ref()
+                    .unwrap()
+                    .clone();
                 let shiny = if *capture.as_event().unwrap().shiny.as_ref().unwrap() {
                     ""
                 } else {
