--- rust-runtime/aws-smithy-http-server/examples/pokemon-service/Cargo.toml
@@ -1,62 +1,0 @@
-[package]
-name = "pokemon-service"
-version = "0.1.0"
-edition = "2021"
-publish = false
-authors = ["Smithy-rs Server Team <smithy-rs-server@amazon.com>"]
-description = "A smithy Rust service to retrieve information about Pokémon."
-default-run = "pokemon-service"
-
-[[bin]]
-name = "pokemon-service"
-path = "src/bin/pokemon-service.rs"
-
-[[bin]]
-name = "pokemon-service-tls"
-path = "src/bin/pokemon-service-tls.rs"
-
-[[bin]]
-name = "pokemon-service-lambda"
-path = "src/bin/pokemon-service-lambda.rs"
-
-[[bin]]
-name = "pokemon-service-connect-info"
-path = "src/bin/pokemon-service-connect-info.rs"
-
-[dependencies]
-async-stream = "0.3"
-clap = { version = "~3.2.1", features = ["derive"] }
-hyper = {version = "0.14.12", features = ["server"] }
-rand = "0.8"
-tokio = "1.20.1"
-tower = "0.4"
-tower-http = { version = "0.3", features = ["trace"] }
-tracing = "0.1"
-tracing-subscriber = { version = "0.3.15", features = ["env-filter", "json"] }
-
-# These dependencies are only required for the `pokemon-service-tls` program.
-tls-listener = { version = "0.5.1", features = ["rustls", "hyper-h2"] }
-tokio-rustls = "0.23.4"
-rustls-pemfile = "1.0.1"
-futures-util = { version = "0.3.16", default-features = false }
-
-# This dependency is only required for the `pokemon-service-lambda` program.
-lambda_http = "0.7.1"
-
-# Local paths
-aws-smithy-http-server = { path = "../../", features = ["aws-lambda", "request-id"] }
-pokemon-service-server-sdk = { path = "../pokemon-service-server-sdk/" }
-
-[dev-dependencies]
-assert_cmd = "2.0"
-home = "0.5"
-serial_test = "0.7.0"
-
-# This dependency is only required for testing the `pokemon-service-tls` program.
-hyper-rustls = { version = "0.23.0", features = ["http2"] }
-
-# Local paths
-aws-smithy-client = { path = "../../../aws-smithy-client/", features = ["rustls"] }
-aws-smithy-http = { path = "../../../aws-smithy-http/" }
-aws-smithy-types = { path = "../../../aws-smithy-types/" }
-pokemon-service-client = { path = "../pokemon-service-client/" }
